Add dashboard components, update packages & CI
Introduce a new dashboard UI (multiple components, hook, utils, types, layout, styles) and remove the legacy pulse page. Bump workspace package versions to 0.2.5 and update dashboard dependencies (add swr, clsx, lucide-react, tailwind-merge, class-variance-authority, etc.). Add a CI job to the preview GitHub workflow to run lint and tests before deploy-preview. Miscellaneous updates: adjust SDK/server engine code and tests, remove e2e-ws scripts, and update package-lock.json to reflect dependency changes.

+++ packages/server/src/engine/inbox.ts
@@ -38,7 +38,7 @@ export async function getInbox(db: Db, workspaceId: string, agentId: string) {
     }
 
     const [result] = await db
-      .select({ count: sql<number>`count(*)::int` })
+      .select({ count: sql<number>`count(*)` })
       .from(messages)
       .where(and(...conditions));
 
@@ -122,7 +122,7 @@ export async function getInbox(db: Db, workspaceId: string, agentId: string) {
     }
 
     const [result] = await db
-      .select({ count: sql<number>`count(*)::int` })
+      .select({ count: sql<number>`count(*)` })
       .from(messages)
       .where(and(...conditions));
